+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,28 @@
 # get-starknet
 
+## 3.3.0
+
+### Minor Changes
+
+- 042997b: Support OKX wallet
+- 042997b: Support MetaMask Starknet snap
+
+### Patch Changes
+
+- Updated dependencies [042997b]
+- Updated dependencies [042997b]
+  - get-starknet-core@3.3.0
+
+## 3.2.1
+
+### Patch Changes
+
+- 9207c26: Support OKX wallet
+- 9207c26: Support MetaMask Starknet snap
+- Updated dependencies [9207c26]
+- Updated dependencies [9207c26]
+  - get-starknet-core@3.2.1
+
 ## 3.0.1
 
 ### Patch Changes